% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_elements_flow.R
\name{add}
\alias{add}
\title{Add a new Node to a Flow (input or function)}
\usage{
add(flow, what = NULL, inputs = NULL, output = NULL, ...)
}
\arguments{
\item{flow}{(a \code{NIflow} object) The flow where to add the new node}

\item{what}{(a function, optional) If present, the function to add as a node to the flow.}

\item{inputs}{(a vector of names, required) The names of new inputs to add to the flow, or the required inputs for the \code{what} function.}

\item{output}{(a character string) The name of the output provided by the \code{what} function.}

\item{...}{named parameters to pass to \code{what}.}
}
\value{
An invisible instance of the flow, with the node added.
}
\description{
Add a new Node to a Flow (input or function)
}
