% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_DLflow.R
\name{DLflow.clone}
\alias{DLflow.clone}
\title{Clone Flow}
\arguments{
\item{deep}{(logical) Deep copy of the flow?, Default: FALSE}
}
\value{
A new flow with exactly the same information.
}
\description{
This function clones a flow.
}
