% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_DLflow.R
\name{DLflow.get_inputs}
\alias{DLflow.get_inputs}
\title{Flow Inputs}
\value{
A character vector with the names of the inputs of the flow.
}
\description{
This function returns the list of names of the inputs to this flow.
}
