% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_DLflow.R
\name{DLflow.get_model}
\alias{DLflow.get_model}
\title{Models from a Flow}
\arguments{
\item{output}{(character) The name of the output that the required model computes.}
}
\value{
A \code{\link{DLmodel}} that is used to compute the given output.
}
\description{
This function returns a model that computes a specific output in a flow.
}
