% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_DLflow.R
\name{DLflow.get_outputs}
\alias{DLflow.get_outputs}
\title{Flow Outputs}
\value{
A character vector with the names of the available outputs.
}
\description{
This function returns the list of names of the outputs available in this flow.
}
