% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_DLflow.R
\name{DLflow.initialize}
\alias{DLflow.initialize}
\title{Flow Creation}
\arguments{
\item{name}{(character) Name of the flow, Default: ''}

\item{inputs}{(list) Character vector or list with the names of the inputs, Default: list()}
}
\value{
A new flow.
}
\description{
This function creates a new \code{DLflow}, with given name and inputs.
}
