% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_DLflow.R
\name{DLflow.replace}
\alias{DLflow.replace}
\title{Replace Part of a Flow}
\arguments{
\item{output}{(character) Name of the output (part of the flow) to substitute.}

\item{with}{(\code{\link{DLmodel}}, \code{\link{DLscheme}} or function) The object that is replacing the previous one.}
}
\value{
The flow with the replaced object.
}
\description{
This function is used to replace a part of the flow, such as a \code{\link{DLmodel}}, a
\code{\link{DLscheme}} or a function, with another object of one these classes.
}
