% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_DLflow.R
\name{DLflow.subset}
\alias{DLflow.subset}
\title{Flow Subsetting}
\arguments{
\item{outputs}{(character vector) List of the outputs to retain.}
}
\value{
A new flow which is complete in the sense that it contains the required \code{outputs} and all parts of the bigger flow
which lead to those outputs.
}
\description{
This function extracts just a subflow from the parent flow, specifying the outputs to retain.
}
