% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_DLflow.R
\name{DLflow.train}
\alias{DLflow.train}
\title{Train Part of a Flow}
\arguments{
\item{output}{(charcater) Name of the output (model) to train.}

\item{input_filenames}{(list) List of input files to train with, that is, the X in the training phase.}

\item{output_filenames}{(list) List of output files, representing the Y in the training phase.}

\item{train_split}{(numeric) Ratio of input files used for training, the remaining are used for validation, Default: 0.75}

\item{epochs}{(numeric) Number of epochs to train for, Default: 10}

\item{target_windows_per_file}{(numeric) Number of windows to extract, as a minimum, per file, Default: 1024}

\item{mode}{(character) Inference mode for the required inputs of this part of the flow, in case they had to be computed, Default: c("debug", "faster", "medium", "slower")}
}
\value{
The flow with the trained model.
}
\description{
This function is used to train a \code{\link{DLmodel}} that specifies an output of the flow.
}
\details{
In case the part of the flow is currently defined by a \code{\link{DLscheme}}, a \code{\link{DLmodel}} is instatiated
taking into account given inputs and output filenames. This is the model to be trained.
}
