% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_elements_flow.R
\name{.add_process}
\alias{.add_process}
\title{Add a Function to a Flow}
\usage{
.add_process(flow, proc, inputs = ifelse(inherits(proc, "function"),
  list(names(formals(proc))), list()), output, ...)
}
\arguments{
\item{flow}{(a flow) The flow where to add a new function}

\item{proc}{(a function) The function itself to be added}

\item{inputs}{(list) List of the inputs needed for the function to be executed, defaults to the formal arguments of the function,
list())}

\item{output}{(name) The name to assign to the output of the function.}

\item{...}{extra parameters for the function, allows to parameterize the flow.}
}
\value{
The flow with the function added
}
\description{
This function adds another function to a given flow.
}
\seealso{
\code{\link[igraph]{add_vertices}},\code{\link[igraph]{add_edges}}
}
