% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_trainable_model.R
\name{.add_trainable_model}
\alias{.add_trainable_model}
\title{Add a Trainable Model}
\usage{
.add_trainable_model(flow, scheme = NULL, inputs, output,
  subset = NULL, verbose = FALSE)
}
\arguments{
\item{flow}{(a DLflow object) The flow where to add the DLmodel}

\item{scheme}{(a DLscheme object) The DLscheme defining the DLmodel, Default: NULL}

\item{inputs}{(list) List of needed inputs (just names, with or without quotes), Default: list()}

\item{output}{(character) The expected output provided by the DLmodel.}

\item{subset}{(NULL) Subset of the problem (see DLmodel especification), Default: NULL}

\item{verbose}{(logical) verbose output?, Default: TRUE}
}
\value{
The flow with the DLmodel added
}
\description{
This function is used to add a (trainable) DLmodel to the flow.
}
