% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_compute_function.R
\name{.build_compute_function}
\alias{.build_compute_function}
\title{Build a Function to Compute Output from Flow}
\usage{
.build_compute_function(args, output)
}
\arguments{
\item{args}{(character) List of arguments to be used in the function definition.}

\item{output}{(character) One of the flow's outputs, which the function will compute.}
}
\value{
A text that is a function definition. The function will be able to compute the given \code{output} from a flow.
}
\description{
This internal function creates the body of a function
which is able to execute the current flow and obtain the given output.
}
\seealso{
\code{\link[stringr]{str_flatten}}
 \code{\link[styler]{style_text}}
}
