% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load_flow.R
\name{.clone_flow}
\alias{.clone_flow}
\title{Clone a Flow}
\usage{
.clone_flow(flow)
}
\arguments{
\item{flow}{(a DLflow object) The flow to clone}
}
\value{
Another DLflow object which is an exact copy of the given \code{flow}.
}
\description{
This functions allows to (deep) clone a flow
}
