% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow.R
\name{.create_flow}
\alias{.create_flow}
\title{Create a Flow (internal)}
\usage{
.create_flow(name = "", inputs = list())
}
\arguments{
\item{inputs}{(list of names) List of inputs used in the flow, Default: list()}
}
\value{
A NIflow object.
}
\description{
This internal function provides an environment usable from the class NIflow.
}
\seealso{
\code{\link[igraph]{make_empty_graph}},\code{\link[igraph]{add_vertices}}
}
