% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_flow.R
\name{.execute_flow}
\alias{.execute_flow}
\title{Execute a Flow}
\usage{
.execute_flow(flow, inputs = list(), given_inputs = NULL,
  desired_outputs = NULL, initialize_outputs = TRUE, cleanup = TRUE,
  mode = c("debug", "faster", "medium", "slower"), verbose = FALSE)
}
\arguments{
\item{flow}{(a NIflow) The flow to be executed.}

\item{inputs}{(list) Inputs to use, Default: list()}

\item{given_inputs}{(list) Additional inputs, Default: NULL}

\item{desired_outputs}{(character array) List of outputs to compute, Default: NULL}

\item{initialize_outputs}{(logical) initialize outputs?, Default: TRUE}

\item{cleanup}{(logical) perform cleanup of intermediate results?, Default: TRUE}

\item{mode}{(character) Mode of execution (only used to run DLmodels), Default: c("debug", "faster", "medium", "slower")}
}
\value{
A list with one (named) field for each \code{desired_output}.
}
\description{
This function allows to run a pipeline (flow) given some of its inputs.
}
\seealso{
\code{\link[neurobase]{readnii}}
}
