% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dependencies.R
\name{.get_dependencies}
\alias{.get_dependencies}
\title{Get Dependencies of a Function}
\usage{
.get_dependencies(script)
}
\arguments{
\item{script}{(a function) A function to determine its dependencies}
}
\value{
A vector with the package dependencies of the given function.
}
\description{
This function allows to obtain the dependencies of a given function.
}
