% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load_flow.R
\name{.load_flow}
\alias{.load_flow}
\title{Load a Flow}
\usage{
.load_flow(filename, verbose = FALSE)
}
\arguments{
\item{filename}{(character) The path to the file to import.}

\item{verbose}{(logical) print information during process?, Default: FALSE}
}
\value{
A DLflow imported from the file.
}
\description{
This function imports a flow from disk
}
