% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow.R
\name{.plot_flow}
\alias{.plot_flow}
\title{Plot a Flow}
\usage{
.plot_flow(flow, interactive = FALSE)
}
\arguments{
\item{flow}{(a DLflow object) The flow.}

\item{interactive}{(logical) should the plot be interactive?, Default: FALSE}
}
\value{
Nothing
}
\description{
This function plots a flow using \link{igraph} plotting capabilities.
}
\seealso{
\code{\link[threejs]{graphjs}}
 \code{\link[igraph]{layout_with_sugiyama}}
 \code{\link[scales]{alpha}},\code{\link[scales]{hue_pal}}
}
