% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_flow.R
\name{.reset_flow}
\alias{.reset_flow}
\title{Reset Flow DLmodels}
\usage{
.reset_flow(flow, outputs = "all")
}
\arguments{
\item{flow}{(a DLflow object) The flow.}

\item{outputs}{(list) List of the names assigned to the DLmodels to reset.}
}
\value{
The flow with the DLmodels resetted.
}
\description{
This functions resets the specified DLmodels to its original state.
}
