% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_flow.R
\name{.reset_outputs}
\alias{.reset_outputs}
\title{Remove Previously Computed Outputs}
\usage{
.reset_outputs(flow)
}
\arguments{
\item{flow}{(a DLflow object) The flow for which outputs are to be removed.}
}
\value{
(Invisibly) the same flow, without pre-computed outputs.
}
\description{
This function deletes previously computed outputs inside the flow.
}
