% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load_flow.R
\name{.save_flow}
\alias{.save_flow}
\title{Save a Flow}
\usage{
.save_flow(flow, path = tempdir(), file_prefix = flow$name)
}
\arguments{
\item{flow}{(a NIflow object) The flow to save.}

\item{path}{(character) Path where to save the file, Default: tempdir()}

\item{file_prefix}{(character) File name, Default: the name of the \code{flow}}
}
\value{
Invisibly, the name of the output file
}
\description{
This function saves a flow to a file.
}
\seealso{
\code{\link[zip]{zip}}
}
