% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_flow.R
\name{.subset_flow}
\alias{.subset_flow}
\title{Subset a Flow}
\usage{
.subset_flow(flow, outputs)
}
\arguments{
\item{flow}{(a DLflow object) The original flow.}

\item{outputs}{(list) Names of the outputs to keep in the new flow.}
}
\value{
A new DLflow object, a copy of the original \code{flow}, but only the given \code{outputs} are kept, as well as their required inputs. That is, the result is a sub-flow whose final results are the given \code{outputs}.
}
\description{
This function allos to extract sub-flows from a given flow.
}
\seealso{
\code{\link[igraph]{topo_sort}},\code{\link[igraph]{make_empty_graph}},\code{\link[igraph]{add_vertices}},\code{\link[igraph]{add_edges}}
}
