% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_flow.R
\name{.train_output}
\alias{.train_output}
\title{Train a DLmodel inside a Flow}
\usage{
.train_output(flow, output, input_filenames, output_filenames,
  given_input = NULL, train_split = 0.75, epochs = 10,
  target_windows_per_file = 1024, metrics_viewer = FALSE,
  mode = c("debug", "faster", "medium", "slower"), verbose = FALSE)
}
\arguments{
\item{flow}{(name) The flow containing the DLmodel}

\item{output}{(name) The name of the output produced by the DLmodel}

\item{input_filenames}{(list) List of filenames to be used as training / testing input data}

\item{output_filenames}{(list) List of filenames to be used as training / testing output data}

\item{given_input}{(list) List of additional inputs, Default: NULL}

\item{train_split}{(numeric) Percentage of data used to train, Default: 0.75}

\item{epochs}{(numeric) Epochs to train the DLmodel for, Default: 10}

\item{mode}{(character) Type of training, Default: c("debug", "faster", "medium", "slower")}

\item{verbose}{(logical) show informative messages?, Default: FALSE}

\item{max_sub_epochs}{(numeric) MAximum number of subepochs, Default: 5}
}
\value{
The flow with the trained DLmodel.
}
\description{
This functions allows to train a DLmodel inside a given flow.
}
\seealso{
\code{\link[neurobase]{fast_readnii}}
 \code{\link[oro.nifti]{as.nifti}}
 \code{\link[RNifti]{writeNifti}}
 \code{\link[oro.nifti]{as.nifti}}
 \code{\link[RNifti]{writeNifti}}
}
