% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_repo.R
\name{download_repo}
\alias{download_repo}
\title{Download a Repository from GitHub or BitBucket}
\usage{
download_repo(repo_name, destination_folder = normalizePath("~"),
  verbose = FALSE)
}
\arguments{
\item{repo_name}{(character) The name of the repository (username/repo).}

\item{destination_folder}{(character) Path where to download the repository.}

\item{verbose}{(logical) print messages?, Default: FALSE}
}
\value{
The path where the repository has been downloaded.
}
\description{
Download a Repository from GitHub or BitBucket
}
\examples{
\dontrun{
download_repo("neuroimaginador/wf4ni", destination_folder = "~/Documents/wf4ni", verbose = TRUE)
}

}
