% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_flow.R
\name{execute}
\alias{execute}
\title{Execute a Flow}
\usage{
execute(flow, inputs = list(), desired_outputs = NULL,
  initialize_outputs = TRUE, ...)
}
\arguments{
\item{flow}{(a NIflow) The flow to be executed.}

\item{inputs}{(list) Inputs to use, Default: list()}

\item{desired_outputs}{(character array) List of outputs to compute, Default: NULL}

\item{initialize_outputs}{(logical) initialize outputs?, Default: TRUE}

\item{...}{Other arguments passed to subfunctions.}
}
\value{
A list with one (named) field for each \code{desired_output}.
}
\description{
This function allows to run a pipeline (flow) given some of its inputs.
}
\seealso{
\code{\link[neurobase]{readnii}}
}
