% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_atlas}
\alias{get_atlas}
\title{Download a Set of Atlases}
\usage{
get_atlas(atlas_name, force_download = FALSE, verbose = FALSE)
}
\arguments{
\item{atlas_name}{(character) Name of the atlas to load}

\item{force_download}{(logical) Force the download of the atlases?. Default: FALSE.}

\item{verbose}{(logical) Print informative messages?. Default: FALSE.}
}
\value{
The path to the (downloaded, if necessary) atlas
}
\description{
Download a Set of Atlases
}
\examples{
get_atlas("MNI152")

}
