% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_dataset}
\alias{get_dataset}
\title{Download a Dataset}
\usage{
get_dataset(dataset, force_download = FALSE, verbose = FALSE)
}
\arguments{
\item{dataset}{(character) Name of the dataset}

\item{force_download}{(logical) Force the download of the dataset?. Default: FALSE.}

\item{verbose}{(logical) Print informative messages?. Default: FALSE.}
}
\value{
The path to the (downloaded, if necessary) dataset
}
\description{
Download a Dataset
}
\examples{
get_dataset("brain_extraction")

}
