% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_flow.R
\name{get_flow}
\alias{get_flow}
\title{Download and Import a Flow}
\usage{
get_flow(repo_name, verbose = FALSE)
}
\arguments{
\item{repo_name}{(character) The name of the repository that hosts the flow.}

\item{verbose}{(logical) print messages?, Default: FALSE}
}
\value{
The imported flow.
}
\description{
This function is used to download a flow from a repository and load it from the local storage. Only GitHub and Bitbucket are currently supported.
}
