% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_flow.R
\name{get_flows_dir}
\alias{get_flows_dir}
\title{Get Location of Installed Flows}
\usage{
get_flows_dir()
}
\value{
The path where flows are installed.
}
\description{
This function returns the path where flows are being installed.
}
