% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge}
\alias{merge}
\title{Merge Two Flows}
\usage{
merge(flow1, flow2, rename = TRUE)
}
\arguments{
\item{flow1}{(NIflow object) First flow to merge}

\item{flow2}{(NIflow object) Second flow to merge}

\item{rename}{(logical) Should inputs/outputs be renamed if the other flow has the same?. Default: TRUE.}
}
\value{
A \code{NIflow} that merges the two flows.
}
\description{
Merge Two Flows
}
