% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_NIflow.R
\name{NIflow.add}
\alias{NIflow.add}
\title{Add an Item to a Flow}
\arguments{
\item{what}{(character, or function) Item to add, Default: NULL}

\item{inputs}{(character) List of required inputs, Default: NULL}

\item{output}{(character) The "name" of the item inside the flow, Default: NULL}
}
\value{
The updated flow.
}
\description{
This function adds an item (an input or a function) to the flow.
}
\details{
If the item to add is an input, the only required (and used) argument is \code{what}. If it's a function, mandatory arguments are \code{what}, \code{inputs} and \code{output}.
}
