% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_NIflow.R
\name{NIflow.check_dependencies}
\alias{NIflow.check_dependencies}
\title{Check Dependencies}
\value{
A logical indicating if all packages are installed or not.
}
\description{
This function checks that all the mandatory packages needed to execute the flow are installed locally.
}
