% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_NIflow.R
\name{NIflow.get_dependencies}
\alias{NIflow.get_dependencies}
\title{Dependencies of a Flow}
\value{
A list with the names of the needed packages.
}
\description{
This function provides the list of packages that are needed in order to execute a flow.#'
}
