% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_NIflow.R
\name{NIflow.get_private}
\alias{NIflow.get_private}
\title{Private Environment of the Flow}
\value{
An environment with the flow internals
}
\description{
This function returns the private environment of the flow
}
