% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_NIflow.R
\name{NIflow.get_process}
\alias{NIflow.get_process}
\title{Processes from a Flow}
\arguments{
\item{output}{(character) The name of the output that the function computes.}
}
\value{
A function that is used to compute the given output.
}
\description{
This function returns a function that computes a specific output in a flow.
}
