% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_NIflow.R
\name{NIflow.get_required_inputs}
\alias{NIflow.get_required_inputs}
\title{Required Inputs}
\arguments{
\item{outputs}{(list of names) Outputs for which to retrieve the mandatory inputs.}
}
\value{
A list with one field for each given output, with the name of the required inputs
}
\description{
This function is used to retrieve the necessary inputs that must be provided in order to compute a specific output.
}
