% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_NIflow.R
\name{NIflow.load}
\alias{NIflow.load}
\title{Load a Flow}
\arguments{
\item{filename}{(character) Path to a file where a flow is stored.}
}
\value{
The loaded flow.
}
\description{
This function loads a flow from disk.
}
