% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_NIflow.R
\name{NIflow.log}
\alias{NIflow.log}
\title{Log a Message Interally}
\arguments{
\item{level}{(character) Level for the message, which indicates its importance, being the most trivial "DEBUG" and the most important "ERROR", Default: c("DEBUG", "INFO", "WARNING", "ERROR")}

\item{message}{(character) The message to log, Default: '...'}
}
\description{
This function is used internally by the flow to log some events.
}
