% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_NIflow.R
\name{NIflow.memory_used}
\alias{NIflow.memory_used}
\title{Memory Used}
\value{
The memory size of the flow.
}
\description{
This function computes the amount of memory used by the flow, including its internals.
}
\seealso{
\code{\link[pryr]{object_size}}
}
