% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_NIflow.R
\name{NIflow.plot}
\alias{NIflow.plot}
\title{Plot a Flow}
\arguments{
\item{interactive}{(logical) Use interactive JavaScript visualization?, Default: FALSE}
}
\description{
This function plots the computation graph of the flow.
}
