% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_NIflow.R
\name{NIflow.print_log}
\alias{NIflow.print_log}
\title{Print Log}
\arguments{
\item{level}{(character) The level for which to print related messages. DEBUG prints all information, Default: c("DEBUG", "WARNING", "INFO", "ERROR")}
}
\value{
Prints in the console the required internal log.
}
\description{
With this function, the internal log of the flow is printed.
}
