% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_NIflow.R
\name{NIflow.replace}
\alias{NIflow.replace}
\title{Replace Part of a Flow}
\arguments{
\item{output}{(character) Name of the output (part of the flow) to substitute.}

\item{with}{(function) The object that is replacing the previous one.}
}
\value{
The flow with the replaced object.
}
\description{
This function is used to replace a part of the flow, such as a function, with another object of one this class.
}
