% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_NIflow.R
\name{NIflow.reset}
\alias{NIflow.reset}
\title{Reset Outputs of a Flow}
\arguments{
\item{outputs}{(character) List of computed outputs to remove, Default: 'all'}
}
\description{
This function deletes previous computed results in a flow.
}
