% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_NIflow.R
\name{NIflow.save}
\alias{NIflow.save}
\title{Save a Flow}
\arguments{
\item{path}{(character) Path where to store the flow, Default: tempdir()}

\item{file_prefix}{(character) Filename, Default: the flow name.}
}
\description{
This function saves a flow to disk.
}
