% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_NIflow.R
\name{NIflow.save_log}
\alias{NIflow.save_log}
\title{Save Log}
\arguments{
\item{filename}{(character) Path to the file where to save the text log.}

\item{level}{(character) Which levels to save to disk? DEBUG stands for all, Default: c("DEBUG", "WARNING", "INFO", "ERROR")}
}
\description{
This function is used to save the internal log of the flow.
}
