% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_NIflow.R
\name{NIflow.to_package}
\alias{NIflow.to_package}
\title{Build Package from Flow}
\arguments{
\item{path}{(character) Path where to create a package}

\item{package_name}{(character) Name of the package to create, Default: the name of the current flow}
}
\description{
With this function, an R package is created that includes the current flow and functions to compute all of its possible outputs.
}
\seealso{
\code{\link[devtools]{use_package}}
}
