% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\title{Plot a Flow}
\usage{
plot(flow, to_file = "")
}
\arguments{
\item{flow}{(a NIflow object) The flow.}

\item{to_file}{(character) Path to a file to save the graph. If not present, plots the flow to the screen. Otherwise, it is stored in a PNG file with the given name.}
}
\value{
Nothing
}
\description{
This function plots a flow using \link{igraph} plotting capabilities.
}
