% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_flow.R
\name{set_flows_dir}
\alias{set_flows_dir}
\title{Set Location to Install Flows to}
\usage{
set_flows_dir(dir)
}
\arguments{
\item{dir}{(character) The new path.}
}
\value{
The new path where to install flows to.
}
\description{
This function allows to change the default location where to install flows.
}
