% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.NIflow}
\alias{summary.NIflow}
\title{Summary of Flow}
\usage{
\method{summary}{NIflow}(flow)
}
\arguments{
\item{flow}{(a \code{NIflow} object) Flow to print summary of.}
}
\value{
Prints inputs, outputs and dependencies between them.
}
\description{
Summary of Flow
}
