% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_names.R
\name{.search_names}
\alias{.search_names}
\title{Search for Variables in Namespaces}
\usage{
.search_names(names, envir = parent.frame())
}
\arguments{
\item{names}{(character) Variable names as strings}

\item{envir}{(environment) Environment where to look for variable names}
}
\value{
If a variable exist in the environment with the corresponding name, its value is returned, otherwise the name is considered as a symbol as is returned as is.
}
\description{
Search for Variables in Namespaces
}
