% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_flow.R
\name{get_flow}
\alias{get_flow}
\title{Download and Import a Flow}
\usage{
get_flow(repo_name, hosts = c("github", "bitbucket"),
  force_download = FALSE)
}
\arguments{
\item{repo_name}{(character) The name of the repository that hosts the flow.}

\item{hosts}{(character) Currently only "github" and "bitbucket" are allowed.}

\item{force_download}{(logical) force the download of the flow, even removing previous versions? Default: FALSE.}
}
\value{
The imported flow.
}
\description{
This function is used to download a flow from a repository and/or load it from the local storage.
}
