% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\title{Plot a Flow}
\usage{
plot(flow, as_interactive = interactive(), to_file = "")
}
\arguments{
\item{flow}{(a NIflow object) The flow.}
}
\value{
Nothing
}
\description{
This function plots a flow using \link{igraph} plotting capabilities.
}
\seealso{
\code{\link[threejs]{graphjs}}
 \code{\link[igraph]{layout_with_sugiyama}}
 \code{\link[scales]{alpha}},\code{\link[scales]{hue_pal}}
}
